Adjusted for Android and New Weapon Overheat Mechanic
In this commit I:
+ Configured the project to run on Android
+ Adjusted
the pause menu to properly fit on a mobile screen
+ Implemented
overheating mechanic so that if you fire continuously for too long the
weapon will jam until it cools down enough
+ Added a UI layer to add a
health bar and a heat bar using Unity sliders.
The .apk contains the
build after these changes.